refactor : 패키지 분리 및 필요없는 클래스 삭제
