--- src/main/java/constants/DateConstants.java
@@ -1,8 +1,0 @@
-package constants;
-
-import domain.Month;
-
-public class DateConstants {
-    public static final int YEAR = 2024;
-    public static final Month MONTH = Month.DECEMBER;
-}
