+++ src/main/java/controller/sub/parent/SubController.java
@@ -1,4 +1,4 @@
-package controller.sub;
+package controller.sub.parent;
 
 public interface SubController {
     void run();
