--- src/main/java/domain/AttendanceCustomDate.java
@@ -1,12 +1,0 @@
-package domain;
-
-import constants.DateConstants;
-
-import java.time.LocalDateTime;
-
-public class AttendanceCustomDate {
-    public static LocalDateTime now() {
-        LocalDateTime now = LocalDateTime.now();
-        return now.withYear(DateConstants.YEAR).withMonth(DateConstants.MONTH.getValue()).withDayOfMonth(13);
-    }
-}
