--- src/main/java/domain/CustomDate.java
@@ -1,7 +1,0 @@
-package domain;
-
-import java.time.LocalDateTime;
-
-public interface CustomDate {
-    LocalDateTime now();
-}
