+++ src/main/java/domain/attendance/AttendanceStatus.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.attendance;
 
 import java.time.DayOfWeek;
 import java.time.LocalDateTime;
