+++ src/main/java/domain/crew/CrewStatus.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.crew;
 
 import exception.InvalidAbsenceCountException;
