+++ src/main/java/exception/CrewNotExistException.java
@@ -1,5 +1,7 @@
 package exception;
 
+import exception.parent.CustomException;
+
 public class CrewNotExistException extends CustomException {
     public CrewNotExistException() {
         super("존재하지 않는 크루입니다.");
