+++ src/main/java/exception/DuplicateAttendanceException.java
@@ -1,5 +1,7 @@
 package exception;
 
+import exception.parent.CustomException;
+
 public class DuplicateAttendanceException extends CustomException {
     public DuplicateAttendanceException() {
         super("이미 출석한 날짜입니다.");
