+++ src/main/java/exception/InvalidAbsenceCountException.java
@@ -1,5 +1,7 @@
 package exception;
 
+import exception.parent.CustomException;
+
 public class InvalidAbsenceCountException extends CustomException {
     public InvalidAbsenceCountException() {
         super("결석 횟수가 올바르지 않습니다.");
