+++ src/main/java/exception/InvalidMenuException.java
@@ -1,5 +1,7 @@
 package exception;
 
+import exception.parent.CustomException;
+
 public class InvalidMenuException extends CustomException {
     public InvalidMenuException() {
         super("기능 번호가 올바르지 않습니다.");
