+++ src/test/java/service/AttendanceModifyServiceTest.java
@@ -1,8 +1,8 @@
 package service;
 
-import constants.DateConstants;
-import domain.AttendanceStatus;
-import domain.Crew;
+import domain.attendance.AttendanceStatus;
+import domain.crew.Crew;
+import domain.date.AttendanceCustomDate;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -32,8 +32,8 @@ class AttendanceModifyServiceTest {
     void test() {
         //given
         LocalDateTime before = LocalDateTime.of(
-                DateConstants.YEAR,
-                DateConstants.MONTH.getValue(),
+                AttendanceCustomDate.YEAR,
+                AttendanceCustomDate.MONTH.getValue(),
                 19,
                 10,
                 30
@@ -44,7 +44,7 @@ class AttendanceModifyServiceTest {
         int date = 19;
         int hour = 10;
         int minutes = 0;
-        LocalDateTime after = LocalDateTime.of(DateConstants.YEAR, DateConstants.MONTH.getValue(), date, hour, minutes);
+        LocalDateTime after = LocalDateTime.of(AttendanceCustomDate.YEAR, AttendanceCustomDate.MONTH.getValue(), date, hour, minutes);
 
         AttendanceModifyResponse response = attendanceModifyService.modify(name, date, hour, minutes);
